// frontend/src/modules/Receipts/types/receiptsType.ts
export interface receiptsType {
    recibe_id: number;
    nombre: string;
    phone: string;
    apt: string;
    mora: boolean;
    days: boolean;
    payment: number;
    dolarBCV: number;
    cedula: string;
    referencia: string;
    porcent: number;
    factura: string;
    date: string;
    user: string,
    accepted: boolean;
}